players can no longer accidentally create mimics

+++ code/game/objects/structures.dm
@@ -45,7 +45,7 @@
 			return
 
 /obj/structure/New()
-	if(prob(0.85)) //Relatively high since windows and tables aren't eligible
+	if(!usr && prob(0.85)) //Relatively high since windows and tables aren't eligible
 		if(!is_type_in_list(src, protected_objects))
 			new /mob/living/simple_animal/hostile/mimic(src.loc, src, null)
 			qdel(src)
